+++ lib/flutter_onscreen_logger.dart
@@ -1,5 +1,6 @@
 library flutter_onscreen_logger;
 
 export 'src/flutter_onscreen_logger.dart';
+export 'src/onscreen_log.dart';
 export 'src/logger_overlay/views/logger_overlay_view.dart';
 export 'src/data/index.dart';