+++ lib/src/data/log_item_type.dart
@@ -1,6 +1,14 @@
+/// Represents the type of log item in the logger.
 enum LogItemType {
+  /// Informational messages
   info,
+
+  /// Success messages
   success,
+
+  /// Warning messages
   warning,
+
+  /// Error messages
   error,
 }